Rename `AdBlockingRecoveryWidget` to `AdBlockingRecoverySetupCTAWidget`.
